+++ src/App.js
@@ -4,9 +4,11 @@ import Button from '@material-ui/core/Button';
 import ButtonGroup from '@material-ui/core/ButtonGroup';
 import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
 import MuiTheme from './MuiTheme.js';
-import Browse from './Browse.js';
-import Shop from './Shop.js';
-import About from './About.js';
+import Browse from './Pages/Browse.js';
+import Shop from './Pages/Shop.js';
+import About from './Pages/About.js';
+import Track from './Pages/Track.js';
+
 import {
   BrowserRouter as Router,
   Switch,
@@ -24,6 +26,7 @@ function App() {
             <ButtonGroup size="large" variant="text" color="primary" aria-label="text primary button group">
               <Button component={Link} to="/browse">Browse</Button>
               <Button component={Link} to="/shop">Shop</Button>
+              <Button component={Link} to="/tracker">Tracker</Button>
               <Button component={Link} to="/about">About</Button>
             </ButtonGroup>
             <div className="body-plate">
@@ -37,6 +40,9 @@ function App() {
                   </Route>
                   <Route path="/shop">
                     <Shop />
+                  </Route>
+                  <Route path="/tracker">
+                    <Track />
                   </Route>
                   <Route path="/">
                     <Browse />
