+++ src/Pages/About.js
@@ -1,16 +1,16 @@
 import React from 'react';
-import './styles.scss';
+import '../styles.scss';
 import Button from '@material-ui/core/Button';
 import ButtonGroup from '@material-ui/core/ButtonGroup';
 import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
-import MuiTheme from './MuiTheme.js';
+import MuiTheme from '../MuiTheme.js';
 
-function Shop() {
+function About() {
   return (
     <MuiThemeProvider theme={MuiTheme}>
       <div className="app" >
               <p>
-                Welcome to the Shop!
+                Welcome to About!
               </p>
               <Button target="_blank"
                 rel="noopener noreferrer" color="secondary">hello</Button>
@@ -27,4 +27,4 @@ function Shop() {
   );
 }
 
-export default Shop;
+export default About;
