+++ src/Pages/index.js
@@ -1,0 +1,1 @@
+export { About, Browse, Shop, Track };
